+++ bots/messageTypes/postback/xdsl.js
@@ -6,7 +6,7 @@ const utils = require("../../utils");
 const Bluebird = require("bluebird");
 const xDSLDiag = require("../../../diagnostics/xdsl");
 const responsesCst = require("../../../constants/responses").FR;
-const v = require("voca");
+const { sprintf } = require("voca");
 
 Bluebird.config({
   warnings: false
@@ -105,7 +105,7 @@ module.exports = [
         .map((offer) => ovhClient.requestPromised("GET", `/xdsl/${offer}`)
           .then((xdslInfo) => new Button(BUTTON_TYPE.POSTBACK, `XDSL_SELECTED_${xdslInfo.accessName}`, xdslInfo.description))
         )
-        .then((buttons) => ({ responses: [createPostBackList(v.sprintf(responsesCst.xdslSelect, Math.floor(1 + (currentIndex / MAX_LIMIT)), Math.ceil(buttons.length / MAX_LIMIT)), buttons, "MORE_XDSL", currentIndex, MAX_LIMIT)], feedback: false }))
+        .then((buttons) => ({ responses: [createPostBackList(sprintf(responsesCst.xdslSelect, Math.floor(1 + (currentIndex / MAX_LIMIT)), Math.ceil(buttons.length / MAX_LIMIT)), buttons, "MORE_XDSL", currentIndex, MAX_LIMIT)], feedback: false }))
         .catch((err) => {
           res.logger.error(err);
           return Bluebird.reject(error(err));