+++ diagnostics/xdsl.js
@@ -1,13 +1,13 @@
 const { TextMessage, Button, ButtonsMessage } = require("../platforms/generics");
 const diagCst = require("../constants/diagnostics").xdsl.FR;
-const v = require("voca");
+const { sprintf } = require("voca");
 
 
 const checkxDSLDiagAdvanced = (diag) => {
   let responses = "";
   let linePb = false;
 
-  responses = v.sprintf(diagCst.diagnosticTime, new Date(diag.diagnosticTime).toUTCString());
+  responses = sprintf(diagCst.diagnosticTime, new Date(diag.diagnosticTime).toUTCString());
 
   if (diag.isModemConnected != null) {
     if (!diag.isModemConnected) {
@@ -23,7 +23,7 @@ const checkxDSLDiagAdvanced = (diag) => {
     for (let i = 0; i < diag.lineDetails.length; i++) {
       // detect sync
       const lineDiag = diag.lineDetails[i];
-      responses += v.sprintf(diagCst.diagnosticLineSync, lineDiag.number, lineDiag.sync ? diagCst.sync : diagCst.unsync);
+      responses += sprintf(diagCst.diagnosticLineSync, lineDiag.number, lineDiag.sync ? diagCst.sync : diagCst.unsync);
       if (lineDiag.lineTest != null) {
         switch (lineDiag.lineTest) {
         case "customerSideProblem":
@@ -70,7 +70,7 @@ const checkxDSLDiag = (xdslOffer, serviceInfos, orderFollowUp, incident, diag) =
   let orderString = "";
 
   if (incident != null) {
-    responses = [new TextMessage(v.sprintf(diagCst.incident, incident.comment || "N/A", incident.endDate || "N/A", `http://travaux.ovh.net/?do=details&id=${incident.taskId}`))];
+    responses = [new TextMessage(sprintf(diagCst.incident, incident.comment || "N/A", incident.endDate || "N/A", `http://travaux.ovh.net/?do=details&id=${incident.taskId}`))];
   }
 
   for (let i = 0; i < orderFollowUp.length; i++) {
@@ -79,7 +79,7 @@ const checkxDSLDiag = (xdslOffer, serviceInfos, orderFollowUp, incident, diag) =
     case "doing":
     case "todo":
     case "error":
-      orderString += v.sprintf(diagCst.orderStepStatus, orderStep.name, diagCst[orderStep.status], `${orderStep.expectedDuration} ${orderStep.durationUnit}`);
+      orderString += sprintf(diagCst.orderStepStatus, orderStep.name, diagCst[orderStep.status], `${orderStep.expectedDuration} ${orderStep.durationUnit}`);
       break;
     case "done":
       if (orderStep.name === "accessIsOperational") {
@@ -92,7 +92,7 @@ const checkxDSLDiag = (xdslOffer, serviceInfos, orderFollowUp, incident, diag) =
   }
 
   if (!orderOk) {
-    return [...responses, new TextMessage(v.sprintf(diagCst.orderNotReady, orderString))];
+    return [...responses, new TextMessage(sprintf(diagCst.orderNotReady, orderString))];
   }
 
   if (xdslOffer.status === "slamming") {
@@ -106,7 +106,7 @@ const checkxDSLDiag = (xdslOffer, serviceInfos, orderFollowUp, incident, diag) =
 
   if (responses.length === 0) {
     const button = new Button("postback", `XDSL_DIAG_${xdslOffer.accessName}`, diagCst.launchDiag);
-    responses = [new ButtonsMessage(`${diagCst.resultOk}\n${v.sprintf(diagCst.resultDiagRemaining, diag ? diag.remaining : 5)}\n${diagCst.resultAdvancedDiag}`, [button])];
+    responses = [new ButtonsMessage(`${diagCst.resultOk}\n${sprintf(diagCst.resultDiagRemaining, diag ? diag.remaining : 5)}\n${diagCst.resultAdvancedDiag}`, [button])];
   }
 
   if (diag) {